separated DefaultLoanToken, DefaultLPS and DefaultToken component
